feat: 링크 목록 조회 영속성 로직 구현

+++ src/main/java/com/seong/shoutlink/domain/link/service/response/FindLinksResponse.java
@@ -4,10 +4,10 @@ import com.seong.shoutlink.domain.link.Link;
 import java.util.List;
 
 public record FindLinksResponse(List<FindLinkResponse> links,
-                                int totalElements,
+                                long totalElements,
                                 boolean hasNext) {
 
-    public static FindLinksResponse of(List<Link> links, int totalElements, boolean hasNext) {
+    public static FindLinksResponse of(List<Link> links, long totalElements, boolean hasNext) {
         List<FindLinkResponse> content = links.stream()
             .map(FindLinkResponse::from)
             .toList();
