+++ src/main/java/com/seong/shoutlink/domain/link/service/result/LinkPaginationResult.java
@@ -3,6 +3,6 @@ package com.seong.shoutlink.domain.link.service.result;
 import com.seong.shoutlink.domain.link.Link;
 import java.util.List;
 
-public record LinkPaginationResult(List<Link> links, int totalElements, boolean hasNext) {
+public record LinkPaginationResult(List<Link> links, long totalElements, boolean hasNext) {
 
 }
